requirements.txt
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.8.1
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Django==5.0.3
django-classy-tags==4.1.0
django-crontab==0.7.1
django-filter==24.2
django-restart==0.2.8.7
djangorestframework==3.15.1
dynaconf==3.2.5
editdistance==0.8.1
feedparser==6.0.11
hvac==2.1.0
idna==3.6
Jinja2==3.1.3
knox==0.1.14
langcodes==3.3.0
loguru==0.7.2
MarkupSafe==2.1.5
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
packaging==24.0
preshed==3.0.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scipy==1.13.0
sgmllib3k==1.0.0
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy-ke==0.1.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
validators==0.24.0
wasabi==1.1.2
weasel==0.3.4
